+++ NEWS.md
@@ -7,6 +7,7 @@
 - `join_coalesce()` coerces x/y args to data.frame and will return a data.frame
 - `most_recent_season()` now internally computes the exact day of the season opener (= Thursday after first Monday of September). (#221)
 - Internal changes to `load_espn_qbr()` to align with nflverse infrastructure design. Also dropped the argument "league" which used to allow loading of College QBR. That's outside of the nflverse scope. (#222)
+- Add new function `stat_mode()`, a re-export from nflfastR, which computes the statistical mode of a vector. (#224)
 
 ---